requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
packaging==24.2
pillow==11.1.0
setuptools==78.1.0
wheel==0.45.1
